Token 유효성 검사 추가
1. Answer
2. Comment
3. Dictionary
4. DummyImage
5. Fcm
6. FolderShare
7. Inquiry
8. Notice
9. User

+++ Spring/Globa/src/main/java/org/y2k2/globa/controller/CommentController.java
@@ -8,19 +8,23 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import org.y2k2.globa.dto.*;
+import org.y2k2.globa.exception.BadRequestException;
 import org.y2k2.globa.service.CommentService;
+import org.y2k2.globa.util.JwtTokenProvider;
 
 import java.net.URI;
 
 @RestController
+@RequestMapping("/folder/{folderId}/record/{recordId}")
 @ResponseBody
 @RequiredArgsConstructor
 public class CommentController {
-    private final String PRE_FIX = "/folder/{folderId}/record/{recordId}";
     private final CommentService commentService;
+    private final JwtTokenProvider jwtTokenProvider;
 
-    @GetMapping(value = PRE_FIX + "/section/{sectionId}/highlight/{highlightId}/comment")
+    @GetMapping(value = "/section/{sectionId}/highlight/{highlightId}/comment")
     public ResponseEntity<?> getComments(
+            @RequestHeader(value = "Authorization") String accessToken,
             @PathVariable("folderId") long folderId,
             @PathVariable("recordId") long recordId,
             @PathVariable("sectionId") long sectionId,
@@ -28,15 +32,19 @@ public class CommentController {
             @RequestParam(required = false, defaultValue = "1", value = "page") int page,
             @RequestParam(required = false, defaultValue = "10", value = "count") int count
     ) {
-        // token 체크
+        if (accessToken == null) {
+            throw new BadRequestException("You must be requested to access token.");
+        }
+        long userId = jwtTokenProvider.getUserIdByAccessToken(accessToken);
 
-        RequestCommentWithIdsDto request = new RequestCommentWithIdsDto(1L, folderId, recordId, sectionId, highlightId);
+        RequestCommentWithIdsDto request = new RequestCommentWithIdsDto(userId, folderId, recordId, sectionId, highlightId);
         ResponseCommentDto commentDto = commentService.getComments(request, page, count);
         return ResponseEntity.ok().body(commentDto);
     }
 
-    @GetMapping(value = PRE_FIX + "/section/{sectionId}/highlight/{highlightId}/comment/{parentId}")
+    @GetMapping(value = "/section/{sectionId}/highlight/{highlightId}/comment/{parentId}")
     public ResponseEntity<?> getReply(
+            @RequestHeader(value = "Authorization") String accessToken,
             @PathVariable("folderId") long folderId,
             @PathVariable("recordId") long recordId,
             @PathVariable("sectionId") long sectionId,
@@ -45,44 +53,56 @@ public class CommentController {
             @RequestParam(required = false, defaultValue = "1", value = "page") int page,
             @RequestParam(required = false, defaultValue = "10", value = "count") int count
     ) {
-        // token 체크
+        if (accessToken == null) {
+            throw new BadRequestException("You must be requested to access token.");
+        }
+        long userId = jwtTokenProvider.getUserIdByAccessToken(accessToken);
 
-        RequestCommentWithIdsDto request = new RequestCommentWithIdsDto(1L, folderId, recordId, sectionId, highlightId, parentId);
+        RequestCommentWithIdsDto request = new RequestCommentWithIdsDto(userId, folderId, recordId, sectionId, highlightId, parentId);
         ResponseReplyDto replyDto = commentService.getReply(request, page, count);
         return ResponseEntity.ok().body(replyDto);
     }
 
-    @PostMapping(value = PRE_FIX + "/section/{sectionId}")
+    @PostMapping(value = "/section/{sectionId}")
     public ResponseEntity<?> addFirstComment(
+            @RequestHeader(value = "Authorization") String accessToken,
             @PathVariable("folderId") long folderId,
             @PathVariable("recordId") long recordId,
             @PathVariable("sectionId") long sectionId,
             @Valid @RequestBody final RequestFirstCommentDto dto
     ) {
-        // token 체크
+        if (accessToken == null) {
+            throw new BadRequestException("You must be requested to access token.");
+        }
+        long userId = jwtTokenProvider.getUserIdByAccessToken(accessToken);
 
-        RequestCommentWithIdsDto request = new RequestCommentWithIdsDto(1L, folderId, recordId, sectionId);
+        RequestCommentWithIdsDto request = new RequestCommentWithIdsDto(userId, folderId, recordId, sectionId);
         long highlightId = commentService.addFirstComment(request, dto);
         return ResponseEntity.created(URI.create("/folder/" + folderId + "/record/" + recordId + "/highlight/" + highlightId)).build();
     }
 
-    @PostMapping(value = PRE_FIX + "/section/{sectionId}/highlight/{highlightId}/comment")
+    @PostMapping(value = "/section/{sectionId}/highlight/{highlightId}/comment")
     public ResponseEntity<?> addComment(
+            @RequestHeader(value = "Authorization") String accessToken,
             @PathVariable("folderId") long folderId,
             @PathVariable("recordId") long recordId,
             @PathVariable("sectionId") long sectionId,
             @PathVariable("highlightId") long highlightId,
             @Valid @RequestBody final RequestCommentDto dto
     ) {
-        // token 체크
+        if (accessToken == null) {
+            throw new BadRequestException("You must be requested to access token.");
+        }
+        long userId = jwtTokenProvider.getUserIdByAccessToken(accessToken);
 
-        RequestCommentWithIdsDto request = new RequestCommentWithIdsDto(2L, folderId, recordId, sectionId, highlightId);
+        RequestCommentWithIdsDto request = new RequestCommentWithIdsDto(userId, folderId, recordId, sectionId, highlightId);
         commentService.addComment(request, dto);
         return ResponseEntity.created(URI.create("/folder/" + folderId + "/record/" + recordId + "/highlight/" + highlightId)).build();
     }
 
-    @PostMapping(value = PRE_FIX + "/section/{sectionId}/highlight/{highlightId}/comment/{parentId}")
+    @PostMapping(value = "/section/{sectionId}/highlight/{highlightId}/comment/{parentId}")
     public ResponseEntity<?> addReply(
+            @RequestHeader(value = "Authorization") String accessToken,
             @PathVariable("folderId") long folderId,
             @PathVariable("recordId") long recordId,
             @PathVariable("sectionId") long sectionId,
@@ -90,15 +110,19 @@ public class CommentController {
             @PathVariable("parentId") long parentId,
             @Valid @RequestBody final RequestCommentDto dto
     ) {
-        // token 체크
+        if (accessToken == null) {
+            throw new BadRequestException("You must be requested to access token.");
+        }
+        long userId = jwtTokenProvider.getUserIdByAccessToken(accessToken);
 
-        RequestCommentWithIdsDto request = new RequestCommentWithIdsDto(1L, folderId, recordId, sectionId, highlightId, parentId);
+        RequestCommentWithIdsDto request = new RequestCommentWithIdsDto(userId, folderId, recordId, sectionId, highlightId, parentId);
         commentService.addReply(request, dto);
         return ResponseEntity.created(URI.create("/folder/" + folderId + "/record/" + recordId + "/highlight/" + highlightId)).build();
     }
 
-    @PatchMapping(value = PRE_FIX + "/section/{sectionId}/highlight/{highlightId}/comment/{commentId}")
+    @PatchMapping(value = "/section/{sectionId}/highlight/{highlightId}/comment/{commentId}")
     public ResponseEntity<?> updateComment(
+            @RequestHeader(value = "Authorization") String accessToken,
             @PathVariable("folderId") long folderId,
             @PathVariable("recordId") long recordId,
             @PathVariable("sectionId") long sectionId,
@@ -106,23 +130,30 @@ public class CommentController {
             @PathVariable("commentId") long commentId,
             @Valid @RequestBody final RequestCommentDto dto
     ) {
-        // token 체크
+        if (accessToken == null) {
+            throw new BadRequestException("You must be requested to access token.");
+        }
+        long userId = jwtTokenProvider.getUserIdByAccessToken(accessToken);
 
-        RequestCommentWithIdsDto requestCommentWithIdsDto = new RequestCommentWithIdsDto(1L, folderId, recordId, sectionId, highlightId);
+        RequestCommentWithIdsDto requestCommentWithIdsDto = new RequestCommentWithIdsDto(userId, folderId, recordId, sectionId, highlightId);
         commentService.updateComment(requestCommentWithIdsDto, commentId, dto);
         return ResponseEntity.noContent().build();
     }
-    @DeleteMapping(value = PRE_FIX + "/section/{sectionId}/highlight/{highlightId}/comment/{commentId}")
+    @DeleteMapping(value = "/section/{sectionId}/highlight/{highlightId}/comment/{commentId}")
     public ResponseEntity<?> deleteComment(
+            @RequestHeader(value = "Authorization") String accessToken,
             @PathVariable("folderId") long folderId,
             @PathVariable("recordId") long recordId,
             @PathVariable("sectionId") long sectionId,
             @PathVariable("highlightId") long highlightId,
             @PathVariable("commentId") long commentId
     ) {
-        // token 체크
+        if (accessToken == null) {
+            throw new BadRequestException("You must be requested to access token.");
+        }
+        long userId = jwtTokenProvider.getUserIdByAccessToken(accessToken);
 
-        RequestCommentWithIdsDto requestCommentWithIdsDto = new RequestCommentWithIdsDto(1L, folderId, recordId, sectionId, highlightId);
+        RequestCommentWithIdsDto requestCommentWithIdsDto = new RequestCommentWithIdsDto(userId, folderId, recordId, sectionId, highlightId);
         commentService.deleteComment(requestCommentWithIdsDto, commentId);
         return ResponseEntity.noContent().build();
     }
